+++ portable-hash/src/hash_impls/hash_core.rs
@@ -17,7 +17,7 @@ use core::{
         Bound, ControlFlow, Deref, Range, RangeFrom, RangeFull, RangeInclusive, RangeTo,
         RangeToInclusive,
     },
-    panic::Location,
+    // explicitly omitted: panic::Location,
     pin::Pin,
     // TODO: ptr::NonNull, (can we safely hash this?)
     // TODO: sync::atomic, (issues with Ordering stability, gating by available atomics, and what ordering to choose)
@@ -65,6 +65,7 @@ impl<T: PortableHash, E: PortableHash> PortableHash for Result<T, E> {
 impl PortableHash for Ordering {
     #[inline]
     fn portable_hash<H: PortableHasher>(&self, state: &mut H) {
+        // TODO(stabilisation): should this match the enum values?
         match self {
             Ordering::Less => state.write_u8(1),
             Ordering::Equal => state.write_u8(2),
@@ -90,6 +91,7 @@ impl PortableHash for Infallible {
 impl PortableHash for CStr {
     #[inline]
     fn portable_hash<H: PortableHasher>(&self, state: &mut H) {
+        // TODO(stabilisation): this differs from std hashers...
         state.write_bytes(self.to_bytes_with_nul());
     }
 }
@@ -235,15 +237,6 @@ impl<T: PortableHash> PortableHash for RangeToInclusive<T> {
     fn portable_hash<H: PortableHasher>(&self, state: &mut H) {
         state.write_u8(5);
         self.end.portable_hash(state);
-    }
-}
-
-impl PortableHash for Location<'_> {
-    #[inline]
-    fn portable_hash<H: PortableHasher>(&self, state: &mut H) {
-        state.write_str(self.file());
-        state.write_u32(self.line());
-        state.write_u32(self.column());
     }
 }